♻️ Refactor: useProductMapModal 훅 추가 및 ProductMapModal 컴포넌트 리팩토링 #5
